v1.1.3.8
In EXECUTOR one of the boss Dark Troopers was too close to a wall. When the map was mirrored he would no longer spawn. Now when mirrored he will be moved to the center of his sector.

Added /reset switch to reset user preferences (such as saved Dark Forces directory).

Adjusted DF autodetect to also detect (and prefer) Remaster.